Tagging 1.0.3 release
git-svn-id: file:///Users/jond/svngit/svnrepos/shopp/tags/1.0.3@338 846ab242-8549-0410-9a75-dfdbdae9473e

+++ core/model/Asset.php
@@ -83,22 +83,31 @@ class Asset extends DatabaseObject {
 		if ($type == "image") $this->setstorage('image');
 		if ($type == "download") $this->setstorage('download');
 
+		if ($this->storage == "fs")	$this->deletefiles($keys);
+
 		$selection = "";
 		foreach ($keys as $value) 
 			$selection .= ((!empty($selection))?" OR ":"")."{$this->_key}=$value OR src=$value";
 
-		if ($this->storage == "fs") $this->deletefiles($selection);
-
 		$query = "DELETE LOW_PRIORITY FROM $this->_table WHERE $selection";
 		$db->query($query);
 	}
-	
-	function deletefiles ($selection) {
+
+	/** 
+	 * deletefiles ()
+	 * Remove files from the file system only when 1 reference to the file exists
+	 * in file references in the database, otherwise, leave them **/
+	function deletefiles ($keys) {
 		$db =& DB::get();
 		
-		$files = $db->query("SELECT name FROM $this->_table WHERE $selection",AS_ARRAY);
-		foreach ($files as $file) 
-			unlink($this->path.$file->name);
+		$selection = "";
+		foreach ($keys as $value) 
+			$selection .= ((!empty($selection))?" OR ":"")."f.{$this->_key}=$value OR f.src=$value";
+		
+		$files = $db->query("SELECT f.name,count(DISTINCT links.id) AS refs FROM $this->_table AS f LEFT JOIN $this->_table AS links ON f.name=links.name WHERE $selection GROUP BY links.name",AS_ARRAY);
+		foreach ($files as $file)
+			if ($file->refs == 1 && file_exists($this->path.$file->name))
+				unlink($this->path.$file->name);
 
 		return true;
 	}
